Create keyboard-row.py
Link to problem:
https://leetcode.com/problems/keyboard-row/description/
Define 3 strings given in the question.
Loop over the array of words.
Loop over each letter of each word.
For a given word, keep append 1,2,3 in check array according to the letter found i.e. if the letter was from string 1 append 1 and so on.
Once you iterate over the complete word, take the set of check list. If the words are from a single row, the length of set of check list should return 1.
Add the word to the answer if check return 1.
Beats 79.5% of the submissions in time complexity.