Create detect-capital.py
Link to the problem:
https://leetcode.com/problems/detect-capital/description/
Keep 2 counters. 1 for the uppercase letters and 1 for lowercase.
Loop over the word. Keep incrementing the counters.
If all letters are upper case then length of the word will be equal to the uppercase counter. same for the lower case.
If first letter is capital and others are lower, upper counter will be 1 and lower will be equal to the length of word - 1.
return True and False according to the requirements.