Create student-attendence-record-i.py
Link to the problem:
https://leetcode.com/problems/student-attendance-record-i/description/
Loop over the string. If there is an 'A' increment the counter.
As soon as the counter reaches 2, return False.
If you encounter 'L', check the next 2 chars but make sure you don't exceed the string length.
If you find 3 consecutive L's return False.
Beats 75.4% of the total submissions in time complexity.