Create longest-harmonious-subsequence.py
Link to the problem:
https://leetcode.com/problems/longest-harmonious-subsequence/description/
Sort the list.
Create a dictionary with keys as list items and values as their count.
Now loop over the list. If next element is increment of previous element, add their count sums.
Keep checking if the current sum exceeds the max sum
Beats 71.9% of total submissions in time complexity.