Create can-place-flowers.py
Link to the problem
https://leetcode.com/problems/can-place-flowers/description/
For each 0, check if its right and left are 0 as well. This means we can plant there so reduce n by 1.
For the starting and ending check if there are 2 consecutive zeros. reduce n by 1 as we can also plant there as well.
check for the corner cases i.e. when n=0 and when length of garden is 1.
Beats 98.1 percent of the total submissions in time complexity.