Create Missing-Number.py
Create a list of 1 more length which consists of -1 only.
Keep iterating original list. On each value of iteration add 1 on the current value index of -1 list.
Only 1 index will remain in the end with -1. That index is the missing number.
Time complexity is O(n).
Link to the problem:
https://leetcode.com/problems/missing-number/description/